refactor: made an s3 and sftp client for mocking
