+++ server/meca/upload/client/s3.js
@@ -1,0 +1,9 @@
+const AWS = require('aws-sdk')
+const config = require('config')
+
+const S3 = new AWS.S3({
+  ...config.get('meca.s3'),
+  apiVersion: '2006-03-01'
+});
+
+module.exports = S3
